Fix Coverity CIDs 1341584-1341589.
(back-ported from commit open-mpi/ompi@3a9664ac9d473142cc8ae6fcde157c96129cea20)

+++ ompi/mca/coll/tuned/coll_tuned_alltoall.c
@@ -237,7 +237,7 @@ int ompi_coll_tuned_alltoall_intra_bruck(void *sbuf, int scount,
     err = ompi_datatype_type_extent (rdtype, &rext);
     if (err != MPI_SUCCESS) { line = __LINE__; goto err_hndl; }
 
-    span = opal_datatype_span(&sdtype->super, size * scount, &gap);
+    span = opal_datatype_span(&sdtype->super, (int64_t)size * scount, &gap);
 
 #ifdef blahblah
     /* try and SAVE memory by using the data segment hung off
