+++ ompi/mca/coll/tuned/coll_tuned_gather.c
@@ -81,8 +81,8 @@ ompi_coll_tuned_gather_intra_binomial(void *sbuf, int scount,
 
     ompi_datatype_type_extent(sdtype, &sextent);
     ompi_datatype_type_extent(rdtype, &rextent);
-    ssize = opal_datatype_span(&sdtype->super, scount * size, &sgap);
-    rsize = opal_datatype_span(&rdtype->super, rcount * size, &rgap);
+    ssize = opal_datatype_span(&sdtype->super, (int64_t)scount * size, &sgap);
+    rsize = opal_datatype_span(&rdtype->super, (int64_t)rcount * size, &rgap);
 
     vrank = (rank - root + size) % size;
